% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fixed.R
\name{decompose}
\alias{decompose}
\title{Compute the SVD of X and construct an orthogonal matrix U_perp such that U_perp * U = 0.}
\usage{
decompose(X, intercept, randomize)
}
\description{
Compute the SVD of X and construct an orthogonal matrix U_perp such that U_perp * U = 0.
}
\keyword{internal}
