% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_asdp.R
\name{divide.sdp}
\alias{divide.sdp}
\title{Approximate a covariance matrix by a block diagonal matrix with blocks
of approximately equal size using Ward's method for hierarchical clustering}
\usage{
divide.sdp(Sigma, max.size)
}
\description{
Approximate a covariance matrix by a block diagonal matrix with blocks
of approximately equal size using Ward's method for hierarchical clustering
}
\keyword{internal}
