% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_asdp.R
\name{divide_sdp}
\alias{divide_sdp}
\title{Approximate a covariance matrix by a block diagonal matrix with blocks
of approximately equal size using Ward's method for hierarchical clustering}
\usage{
divide_sdp(Sigma, nBlocks = 10)
}
\description{
Approximate a covariance matrix by a block diagonal matrix with blocks
of approximately equal size using Ward's method for hierarchical clustering
}
\keyword{internal}
