% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stats_fs.R
\name{fs}
\alias{fs}
\title{Forward selection}
\usage{
fs(X, y, omp = FALSE)
}
\arguments{
\item{X}{matrix of predictors}

\item{y}{response vector}

\item{omp}{whether to use orthogonal matching pursuit (OMP)}
}
\value{
vector with jth component the variable added at step j
}
\description{
Perform forward variable selection with or without OMP
}
\keyword{internal}

