% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create.R
\name{knockoff.create}
\alias{knockoff.create}
\title{Create knockoff variables}
\usage{
knockoff.create(X, method = c("equicorrelated", "sdp"), randomize = F)
}
\arguments{
\item{X}{normalized n-by-p design matrix (n >= 2p)}

\item{method}{either 'equicorrelated' or 'sdp'}

\item{randomize}{whether the knockoffs are deterministic or randomized}
}
\value{
The n-by-p knockoff matrix
}
\description{
Creates knockoff variables for the original variables.
}
\details{
To use SDP knockoffs, you must have a Python installation with
CVXPY. For more information, see the vignette on SDP knockoffs:

\code{vignette('sdp', package='knockoff')}
}

