% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/knockoff.R
\docType{package}
\name{knockoff}
\alias{knockoff}
\alias{knockoff-package}
\title{The knockoff filter for controlling the false discovery rate}
\description{
The knockoff filter for controlling the false discovery rate
}

