% Generated by roxygen2 (4.0.2): do not edit by hand
\name{knockoff.threshold}
\alias{knockoff.threshold}
\title{Threshold for the knockoff filter}
\usage{
knockoff.threshold(W, fdr = 0.2, method = c("knockoff", "knockoff+"))
}
\arguments{
\item{W}{the test statistics}

\item{fdr}{target FDR (false discovery rate)}

\item{method}{either 'knockoff or 'knockoff+'}
}
\value{
The threshold for variable selection.
}
\description{
Computes the threshold for the knockoff filter.
}

