% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_sqrt_lasso.R
\name{sqrt_lasso_coeffs}
\alias{sqrt_lasso_coeffs}
\title{SQRT lasso}
\usage{
sqrt_lasso_coeffs(X, y, nlambda = NULL, ...)
}
\arguments{
\item{X}{matrix of predictors}

\item{y}{response vector}
}
\value{
vector with jth component the selection probability of variable j
}
\description{
Perform variable selection with SQRT lasso
}
\keyword{internal}
