% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_stability_selection.R
\name{stability_selection_importance}
\alias{stability_selection_importance}
\title{Stability selection}
\usage{
stability_selection_importance(X, y, ...)
}
\arguments{
\item{X}{matrix of predictors}

\item{y}{response vector}
}
\value{
vector with jth component the selection probability of variable j
}
\description{
Perform variable selection with stability selection
}
\keyword{internal}
