% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knockoff_filter.R
\name{verify_stat_depends}
\alias{verify_stat_depends}
\title{Verify dependencies for chosen statistics}
\usage{
verify_stat_depends(statistic)
}
\arguments{
\item{statistic}{the statistic chosen by the user}
}
\description{
Verify dependencies for chosen statistics
}
\keyword{internal}
