% Generated by roxygen2 (4.0.2): do not edit by hand
\name{knockoff.create}
\alias{knockoff.create}
\title{Create knockoff variables}
\usage{
knockoff.create(X, method = c("equicorrelated", "sdp"), randomized = F)
}
\arguments{
\item{X}{normalized n-by-p design matrix (n >= 2p)}

\item{method}{either 'equicorrelated' or 'sdp'}

\item{randomized}{whether the knockoffs are deterministic or randomized}
}
\value{
The n-by-p knockoff matrix
}
\description{
Creates knockoff variables for the original variables.
}

