% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knockoff_filter.R
\name{knockoff.threshold}
\alias{knockoff.threshold}
\title{Threshold for the knockoff filter}
\usage{
knockoff.threshold(W, fdr = 0.1, offset = 1)
}
\arguments{
\item{W}{the test statistics}

\item{fdr}{target false discovery rate (default: 0.1)}

\item{offset}{either 0 or 1 (default: 1). The offset used to compute the rejection threshold on the
statistics. The value 1 yields a slightly more conservative procedure ("knockoffs+") that
controls the FDR according to the usual definition, while an offset of 0 controls a modified FDR.}
}
\value{
The threshold for variable selection.
}
\description{
Computes the threshold for the knockoff filter.
}
