% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lasso_max_lambda_lars}
\alias{lasso_max_lambda}
\alias{lasso_max_lambda_glmnet}
\alias{lasso_max_lambda_lars}
\title{Maximum lambda in lasso model}
\usage{
lasso_max_lambda_lars(X, y)

lasso_max_lambda_glmnet(X, y, nlambda = NULL)

lasso_max_lambda(X, y, method = c("glmnet", "lars"), ...)
}
\arguments{
\item{X}{matrix of predictors}

\item{y}{response vector}
}
\value{
vector of maximum lambda's
}
\description{
Computes the earliest (largest) lambda's for which predictors enter the
lasso model.
}
\keyword{internal}

