% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_glmnet.R
\name{lasso_max_lambda}
\alias{lasso_max_lambda}
\title{Maximum lambda in lasso model}
\usage{
lasso_max_lambda(X, y, method = c("glmnet", "lars"), ...)
}
\arguments{
\item{X}{matrix of predictors}

\item{y}{response vector}

\item{method}{either 'glmnet' or 'lars'}
}
\value{
vector of maximum lambda's
}
\description{
Computes the earliest (largest) lambda's for which predictors enter the
lasso model.
}
\keyword{internal}
